requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
coloredlogs==15.0.1
gitdb==4.0.9
GitPython==3.1.24
humanfriendly==10.0
idna==3.3
lxml==4.6.4
pyreadline3==3.3
requests==2.26.0
smmap==5.0.0
typing-extensions==4.0.1
urllib3==1.26.7
